Add backend and FE url

+++ src/main/java/com/backendsyndicate/smashclub/admin/dto/response/RespAdminPlayerListDTO.java
@@ -11,7 +11,7 @@ public class RespAdminPlayerListDTO {
     private String id;
     private String fullName;
     private String email;
-    private int status;
+    private byte status;
     private String createdAt;
     private String lockedUntil;
 }
